backend/requirements.txt
# Background task scheduler (for automatic itinerary regeneration)
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
celery==5.3.1
redis==4.6.0
numpy==1.26.4
pandas==2.1.2
scikit-learn==1.2.2
joblib==1.3.2
faker==25.0.1
django-q2==1.7.4
pytest
requests
pydantic>=2.0
fastapi>=0.95
boto3
geopy>=2.3.0
geopy==2.3.0
joblib
